Updated for latest header PR
