+++ lib/CMakeLists.txt
@@ -61,6 +61,7 @@ if(LIBOCPP_ENABLE_V201)
             ocpp/v201/database_handler.cpp
             ocpp/v201/device_model.cpp
             ocpp/v201/device_model_storage_sqlite.cpp
+            ocpp/v201/enums.cpp
             ocpp/v201/evse.cpp
             ocpp/v201/evse_manager.cpp
             ocpp/v201/init_device_model_db.cpp
